* Use VkFlags immediately, update git repository URL.

+++ Colorful/Vulkan/VulkanPipeline.cpp
@@ -346,20 +346,20 @@ VulkanRenderPassPtr VkRenderStateCache::getOrCreateRenderPass(const RenderPassDe
 		{
 			VK_SUBPASS_EXTERNAL,
 			0u,
-			static_cast<uint64_t>(RenderPassDesc::EPipelineStageFlags::BottomOfPipe),
-			static_cast<uint64_t>(RenderPassDesc::EPipelineStageFlags::ColorAttachmentOutput),
-			static_cast<uint64_t>(RenderPassDesc::EAccessFlags::MemoryRead),
-			static_cast<uint64_t>(RenderPassDesc::EAccessFlags::ColorAttachmentRead) | static_cast<uint64_t>(RenderPassDesc::EAccessFlags::ColorAttachmentWrite),
+			RenderPassDesc::EPipelineStageFlags::BottomOfPipe,
+			RenderPassDesc::EPipelineStageFlags::ColorAttachmentOutput,
+			RenderPassDesc::EAccessFlags::MemoryRead,
+			RenderPassDesc::EAccessFlags::ColorAttachmentRead | RenderPassDesc::EAccessFlags::ColorAttachmentWrite,
 		};
 
 		defaultRenderPassDesc.SubPassDependencies[1] =
 		{
 			0u,
 			VK_SUBPASS_EXTERNAL,
-			static_cast<uint64_t>(RenderPassDesc::EPipelineStageFlags::ColorAttachmentOutput),
-			static_cast<uint64_t>(RenderPassDesc::EPipelineStageFlags::BottomOfPipe),
-			static_cast<uint64_t>(RenderPassDesc::EAccessFlags::ColorAttachmentRead) | static_cast<uint64_t>(RenderPassDesc::EAccessFlags::ColorAttachmentWrite),
-			static_cast<uint64_t>(RenderPassDesc::EAccessFlags::MemoryRead),
+			RenderPassDesc::EPipelineStageFlags::ColorAttachmentOutput,
+			RenderPassDesc::EPipelineStageFlags::BottomOfPipe,
+			RenderPassDesc::EAccessFlags::ColorAttachmentRead | RenderPassDesc::EAccessFlags::ColorAttachmentWrite,
+			RenderPassDesc::EAccessFlags::MemoryRead,
 		};
 
 		auto renderPass = create_shared<VulkanRenderPass>(defaultRenderPassDesc);
